+++ CHANGELOG.md
@@ -41,6 +41,7 @@ doivent être mises à jour #1667
 - Les utilisateurs avec le droit system:edition peuvent désormais changer manuellement l'état d'un dossier #2354
 - La variable d'environnement JOURNAL_MAX_AGE_IN_MONTHS est remplacée par JOURNAL_MAX_AGE_IN_DAYS (à 60 jours par défaut) #2358
 - Uniformisation des listes de sélection de type de dossier #2351
+- Il est possible de modifier un dossier en état "Erreur lors de la transformation du dossier" #2332
 
 ## Suppressions